feat: fixed Focued Line's Interface
