+++ app/components/Reditor/components/Node/FocuedLine.tsx
@@ -1,62 +1,35 @@
 import React from 'react';
 import styled from '@xstyled/styled-components';
-import { cursorS, useFocus, useNoteOp, _cursor2S } from 'app/models/Cursor';
+import { useFocus3, useNoteOp, _cursor2S } from 'app/models/Cursor';
 import { useRecoilValue } from 'recoil';
 import { useHotKeyMapping } from '../../hooks/useHotKeyMapping';
-import { NoteId } from 'app/models/notes/typings/note';
 import { Indents } from './Indents';
 import { Cursor } from '../Cursor';
-import { textWithIndents } from '../../utils/parsers/parser';
 import { Normal } from './Normal';
-import { LineProps } from './Line';
+import { LineProps } from './ViewLine';
 
 /**
  * TODO:
- * - まずlineを表示
  * - 次にカーソルを表示
  * - 次にカーソルを動かす
  */
 
-type Props = {
-  noteId: NoteId;
-};
-
-export const FocusedLine: React.VFC<LineProps> = ({ value, lineIndex }) => {
-  const cursor = useRecoilValue(cursorS);
+export const FocusedLine: React.VFC<LineProps> = ({ line, lineIndex }) => {
+  // const cursor = useRecoilValue(cursorS);
   const cursor2 = useRecoilValue(_cursor2S);
-  const { ref: textareaRef, onFocus } = useFocus();
-  // const { level, value } = textWithIndents.tryParse(cursor.line?.value ?? '');
+  const { ref: textareaRef } = useFocus3();
   const keys = useNoteOp(0);
-  // const isFocus = noteId === cursor.noteId;
   const { keyMapping } = useHotKeyMapping(true, keys);
 
-  // if (!isFocus) {
-  //   return null;
-  // }
-
   return (
-    <div>
-      <Indents level={0} />
-      <Normal value={'hgoehogehoghoge'} lineIndex={cursor2.pos.ln} />
+    <Wrap>
+      <Indents level={line.indent} />
+      <Normal value={line.nodeValue} lineIndex={cursor2.pos.ln} />
       <Cursor noteId={0} textareaRef={textareaRef} onKeyDown={keyMapping} />
-    </div>
+    </Wrap>
   );
-
-  // return (
-  //   <Span top={cursor.pxPos?.top ?? 0}>
-  //     <Indents level={level} />
-  //     <Normal value={'hgoehogehoghoge'} lineIndex={cursor2.pos.ln} />
-  //     <Cursor
-  //       noteId={noteId}
-  //       textareaRef={textareaRef}
-  //       onKeyDown={keyMapping}
-  //     />
-  //   </Span>
-  // );
 };
 
-const Span = styled.span<{ top: number }>`
-  position: absolute;
-  top: ${p => p.top}px;
+const Wrap = styled.div`
   background-color: #ff8787;
 `;
